elevaite/python_apps/flyte-worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = ""
authors = [
  {name="John Zachariadis", email="<johnbelly95@gmail.com>"}
]
readme = "README.md"
requires-python = "~=3.11"


dependencies = [
"pika>=1.3.2",
"flytekit~=1.12.0",
"pandas>=2.2.2",
"scikit-learn>=1.4.2",
"lakefs>=0.6.0",
"unstructured>=0.13.6",
"qdrant-client>=1.9.1",
"openai>=1.26",
"python-docx>=1.1.2",
"elasticsearch>=8.13.0",
"redis>=5.0.4",
"python-dotenv>=1.0.1",
"tiktoken>=0.6.0",
"boto3>=1.34.69",
"botocore>=1.34.69",
"flytekitplugins-pydantic>=1.12.0",
"flytekitplugins-envd>=1.12.0",
"flyteidl>=1.12.0",
"elevaite_client"
]

[tool.uv.sources]
"elevaite-client" = { workspace = true }
